[Optimization] Use deepcopy to update graph.paths
This solves the following bug: if more than one action is appended at
the end of a 'branch', only the last one is considered if optimization 2
is selected; this happens because in swap_children we used to rely on
the steps as separate object, but they're actually lists referring to
the same objects, so if one of them is modified all of them are too.

This is just TEMPORARY: asap a nicer and more performant solution will
be implemented, possibly with generators.